+++ sqlConsole.py
@@ -1255,7 +1255,7 @@ class resultSet(QTableWidget):
         clr = QColor(cfg('highlightColor', '#def'))
         hlBrush = QBrush(clr)
 
-        clr = QColor(clr.red()*0.9, clr.green()*0.9, clr.blue()*0.95)
+        clr = QColor(int(clr.red()*0.9), int(clr.green()*0.9), int(clr.blue()*0.95))
         hlBrushLOB = QBrush(clr)
         
         wBrush = QBrush(QColor('#ffffff'))
